+++ apps/api-service/src/pairpilot/pairpilot.module.ts
@@ -1,16 +1,16 @@
 import { Module } from "@nestjs/common"
-import { IntelligenceService } from "./intelligence.service"
-import { IntelligenceController } from "./intelligence.controller"
+import { PairPilotService } from "./pairpilot.service"
+import { PairPilotController } from "./pairpilot.controller"
 import { CqrsModule } from "@nestjs/cqrs"
 import { DbConnectionMap } from "src/shared/utils/db-connection.map"
 import { Thread, ThreadSchema } from "./schemas/thread.schema"
 import { CreateThreadCommandHandler } from "./commands/handler/create-thread.handler"
-import { IntelligenceRepository } from "./intelligence.repository"
+import { PairPilotRepository } from "./pairpilot.repository"
 import { EntityModule } from "@/shared/entity/entity.module"
 import { FetchThreadByIdQueryHandler } from "./queries/handler/fetch-thread-by-id.handler"
 import { HttpModule } from "@nestjs/axios"
-import { IntelligenceStrategy } from "./intelligence.strategy"
-import { IntelligenceAgent } from "./intelligence.agent"
+import { PairPilotStrategy } from "./pairpilot.strategy"
+import { PairPilotAgent } from "./pairpilot.agent"
 
 @Module({
   imports: [
@@ -21,14 +21,14 @@ import { IntelligenceAgent } from "./intelligence.agent"
       DbConnectionMap.Primary
     ),
   ],
-  controllers: [IntelligenceController],
+  controllers: [PairPilotController],
   providers: [
-    IntelligenceService,
-    IntelligenceRepository,
-    IntelligenceStrategy,
-    IntelligenceAgent,
+    PairPilotService,
+    PairPilotRepository,
+    PairPilotStrategy,
+    PairPilotAgent,
     CreateThreadCommandHandler,
     FetchThreadByIdQueryHandler,
   ],
 })
-export class IntelligenceModule {}
+export class PairPilotModule {}