+++ apps/api-service/src/app.module.ts
@@ -3,14 +3,14 @@ import { CoreModule } from "./core/core.module"
 import { EventEmitterModule } from "@nestjs/event-emitter"
 import { EntityModule } from "./shared/entity/entity.module"
 import { AppController } from "./app.controller"
-import { PairPilotModule } from "./pairpilot/pairpilot.module"
+import { IntelligenceModule } from "./intelligence/intelligence.module"
 
 @Module({
   imports: [
     EventEmitterModule.forRoot(),
     CoreModule,
     EntityModule,
-    PairPilotModule,
+    IntelligenceModule,
   ],
   controllers: [AppController],
   providers: [],